requirements.txt
python-telegram-bot==13.7
feedparser==6.0.8
date-util==2.8.1
